+++ buildspecs/buildspec.release.yml
@@ -1,0 +1,17 @@
+version: 0.2
+
+env:
+  variables:
+    AWS_REGION: us-west-2
+  parameter-store:
+    NUGET_API_KEY: NuGetAPIKey
+phases:
+  install:
+    runtime-versions:
+      dotnet: 3.1
+  build:
+    commands:
+      - dotnet restore
+      - dotnet pack -c Release --version-suffix "ci-$(date +%s)"
+      - dotnet nuget push ./src/Amazon.CloudWatch.EMF/bin/Release/Amazon.CloudWatch.EMF.*.nupkg -k $NUGET_API_KEY
+      - dotnet nuget push ./src/Amazon.CloudWatch.EMF.Web/bin/Release/Amazon.CloudWatch.EMF.Web.*.nupkg -k $NUGET_API_KEY
