+++ buildspecs/buildspec.yml
@@ -15,7 +15,6 @@ phases:
       # https://docs.aws.amazon.com/codebuild/latest/userguide/sample-docker-custom-image.html#sample-docker-custom-image-files
       - nohup /usr/local/bin/dockerd --host=unix:///var/run/docker.sock --host=tcp://127.0.0.1:2375 --storage-driver=overlay2&
       - timeout 15 sh -c "until docker info; do echo .; sleep 1; done"
-      - npm install
   build:
     commands:
       - pwd && ./scripts/start-agent.sh
