+++ src/Amazon.CloudWatch.EMF/Amazon.CloudWatch.EMF.csproj
@@ -3,7 +3,14 @@
     <PropertyGroup>
         <PackageId>Amazon.CloudWatch.EMF</PackageId>
         <VersionPrefix>0.1.0</VersionPrefix>
-        <Authors>AWS Web Services</Authors>
+        <Authors>Amazon Web Services</Authors>
+        <Description>Amazon CloudWatch Embedded Metric Format Client Library</Description>
+        <Language>en-US</Language>
+        <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
+        <ProjectUrl>https://github.com/awslabs/aws-embedded-metrics-dotnet</ProjectUrl>
+        <RepositoryUrl>https://github.com/awslabs/aws-embedded-metrics-dotnet</RepositoryUrl>
+        <PackageTags>AWS Amazon cloud CloudWatch Metrics EMF</PackageTags>
+        <PackageIconUrl>http://media.amazonwebservices.com/aws_singlebox_01.png</PackageIconUrl>
     </PropertyGroup>
 
     <PropertyGroup>
