+++ tests/Amazon.CloudWatch.EMF.Canary/Dockerfile
