+++ tests/Amazon.CloudWatch.EMF.Canary/Program.cs
