+++ tests/Amazon.CloudWatch.EMF.Canary/container-definitions.json
