+++ tests/Amazon.CloudWatch.EMF.IntegrationTests/Amazon.CloudWatch.EMF.IntegrationTests.csproj
@@ -22,7 +22,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Amazon.CloudWatch.EMF\Amazon.CloudWatch.EMF.csproj" />
+      <ProjectReference Include="..\..\src\Amazon.CloudWatch.EMF\Amazon.CloudWatch.EMF.csproj" />
     </ItemGroup>
 
 </Project>
