+++ tests/Amazon.CloudWatch.EMF.IntegrationTests/MetricsLoggerIntegrationTest.cs
