+++ tests/Amazon.CloudWatch.EMF.Tests/Amazon.CloudWatch.EMF.Tests.csproj
@@ -28,7 +28,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Amazon.CloudWatch.EMF\Amazon.CloudWatch.EMF.csproj" />
+      <ProjectReference Include="..\..\src\Amazon.CloudWatch.EMF\Amazon.CloudWatch.EMF.csproj" />
     </ItemGroup>
 
 </Project>
