+++ tests/Amazon.CloudWatch.EMF.Tests/Environment/DefaultEnvironmentTests.cs
