+++ tests/Amazon.CloudWatch.EMF.Tests/Environment/EC2EnvironmentTests.cs
