+++ tests/Amazon.CloudWatch.EMF.Tests/Environment/EnvironmentProviderTests.cs
