+++ tests/Amazon.CloudWatch.EMF.Tests/Environment/LambdaEnvironmentTests.cs
