+++ tests/Amazon.CloudWatch.EMF.Tests/Logger/MetricsLoggerTests.cs
