+++ tests/Amazon.CloudWatch.EMF.Tests/Model/MetricDefinitionTests.cs
