+++ tests/Amazon.CloudWatch.EMF.Tests/Model/MetricDirectiveTests.cs
