+++ tests/Amazon.CloudWatch.EMF.Tests/Sink/AgentSinkTests.cs
