+++ tests/Amazon.CloudWatch.EMF.Tests/Sink/EndpointTests.cs
