+++ tests/Amazon.CloudWatch.EMF.Tests/Sink/MockSink.cs
